Organized some of the layout names/ids for better clarity.
Added tour functionality. When a user clicks on tours, it brings them to a list of tours available to take. Now when you select a tour, that will take you to a list of sites available in that tour (in order). When you click on the site, it will fire up gmaps with current loc and destination of the site.